requirements.txt
httpx>=0.27.0
websockets>=12.0
yfinance>=0.2.43

numpy>=1.26.0
pandas>=2.2.0
scipy>=1.13.0
matplotlib>=3.8.0
streamlit>=1.37.0

pytest>=8.0.0
pytest-asyncio>=0.23.0

black>=24.0.0
flake8>=7.0.0
mypy>=1.8.0

python-dotenv>=1.0.1
